# docs/openapi.yaml
openapi: 3.0.3
info:
  title: psauerborn.dev API
  version: "1.0.0"
  description: |
    Public API backing psauerborn.dev. It serves the CV, the article catalogue
    and article content, the subagent and spec catalogues, the project listing,
    and accepts contact messages and article comments.

    Conventions:

    - Every endpoint is served under the `/v1` prefix.
    - Requests and responses are `application/json`, except the two content
      endpoints, which return raw bytes as `binary/octet-stream`.
    - Every error response (400, 404, 500) uses the shared `Error` envelope.
    - The API is entirely public: no endpoint requires authentication, and no
      security scheme is defined.

    CORS: cross-origin requests are accepted from `http://localhost:9000`,
    `https://psauerborn.dev` and `https://dev.psauerborn.dev`, for the methods
    GET, POST, PATCH, PUT, DELETE and OPTIONS, with the `Content-Type`,
    `Accept`, `Accept-Language` and `Content-Language` request headers.
    Preflight `OPTIONS` requests are answered with `204 No Content` and an empty
    body for every origin; only allowed origins receive the
    `Access-Control-Allow-*` response headers, so browsers reject the response
    for any other origin. The preflight handling applies to every path below and
    is therefore not repeated per operation.
servers:
  - url: "{scheme}://{host}"
    description: API base URL
    variables:
      scheme:
        default: https
        enum:
          - http
...
paths:
...
  /v1/messages:
    post:
      tags: [messages]
      operationId: createMessage
      summary: Submit a contact message
      description: >-
        Records a contact message and returns the identifier generated for it.
        The email is trimmed and lower-cased. The name and organization are
        trimmed, runs of internal whitespace are collapsed to a single space and
        the first letter of each word is capitalized, while the rest of each
        word is stored exactly as submitted.
        The body is bound tolerantly: fields beyond the documented schema are
        ignored rather than rejected. The endpoint is also served on the
        trailing-slash path `/v1/messages/`, which is identical in every
        respect.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateMessageRequest"
      responses:
        "201":
          description: The message was recorded.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CreateMessageResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "500":
          $ref: "#/components/responses/InternalServerError"
...
  /v1/articles/{article_id}/content:
    get:
      tags: [articles]
      operationId: getArticleContent
      summary: Return the raw content of an article
      description: >-
        Returns the raw bytes of the document linked to the article, written as
        `binary/octet-stream` rather than as JSON. An article that does not
        exist, is hidden, has no linked document, or whose document is
        restricted is answered with the same 404 envelope.
      parameters:
        - $ref: "#/components/parameters/ArticleID"
      responses:
        "200":
          description: The raw article content.
          content:
            binary/octet-stream:
              schema:
                type: string
                format: binary
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalServerError"
...
  /v1/articles/{article_id}/comment:
    post:
      tags: [articles]
      operationId: createArticleComment
      summary: Comment on an article
      description: >-
        Records the submitted comment against the article and returns the
        identifier generated for it. The author is trimmed, runs of internal
        whitespace are collapsed to a single space and the first letter of each
        word is capitalized, while the rest of each word is stored exactly as
        submitted; a comment submitted without an author is stored anonymously.
        The body is bound tolerantly: fields beyond the documented schema are
        ignored rather than rejected. An article that is not publicly visible
        yields the same 404 envelope as the read endpoints.
      parameters:
        - $ref: "#/components/parameters/ArticleID"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateCommentRequest"
      responses:
        "201":
          description: The comment was recorded.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CreateCommentResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalServerError"
...
components:
  parameters:
    ArticleID:
      name: article_id
      in: path
      required: true
      description: Identifier of the article.
      schema:
        type: string
      example: 8f1f1f52-3f1e-4a34-9f1e-1b0f1e2d3c4b
...
  responses:
    BadRequest:
      description: >-
        The request body could not be parsed, or a field failed validation. The
        `details` value names the offending field and the reason it was
        rejected.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
          example:
            error: Bad Request
            details: "comment: must not be empty"
    NotFound:
      description: >-
        The requested resource is not available. The same envelope is returned
        whether the resource does not exist, is hidden, has no linked document
        or is restricted, so that the cases cannot be told apart.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
          example:
            error: Not Found
            details: The requested resource could not be found.
    InternalServerError:
      description: >-
        An unhandled error occurred. The envelope is fixed: the underlying
        failure is logged for operators and never disclosed to the caller.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
          example:
            error: Internal Server Error
            details: Something went wrong.
  schemas:
    Error:
      type: object
      description: The envelope returned by every error response of this API.
      required:
        - error
        - details
      properties:
        error:
          type: string
          description: >-
            Generic error code mirroring the HTTP status reason phrase of the
            response it accompanies.
          enum:
            - Bad Request
            - Not Found
            - Internal Server Error
        details:
          type: string
          description: >-
            Additional context that is safe to disclose. Validation failures
            carry `<field>: <reason>`; all other errors carry a fixed, generic
            string.
          example: "comment: must not be empty"
...
    CreateMessageRequest:
      type: object
      description: >-
        Contact message to record. Fields beyond this schema are ignored rather
        than rejected.
      required:
        - email
        - name
        - message
      properties:
        email:
          type: string
          format: email
          description: >-
            Email address of the sender. It is trimmed and lower-cased before it
            is stored; an invalid or over-long value yields a 400.
        name:
          type: string
          description: >-
            Name of the sender. It is trimmed, runs of internal whitespace are
            collapsed to a single space and the first letter of each word is
            capitalized, while the rest of each word is stored exactly as
            submitted; an empty or over-long value yields a 400.
        organization:
          type: string
          description: >-
            Organization of the sender. Optional: an absent or empty value is
            stored as null. An over-long value yields a 400.
        message:
          type: string
          description: >-
            Message body, stored as submitted. An empty or whitespace-only value
            yields a 400.
    CreateMessageResponse:
      type: object
      required:
        - message_id
      properties:
        message_id:
          type: string
          description: Identifier generated for the recorded message.
...
    CreateCommentRequest:
      type: object
      description: >-
        Comment to record against the article. Fields beyond this schema are
        ignored rather than rejected.
      required:
        - comment
      properties:
        author:
          type: string
          description: >-
            Author of the comment. Optional: an absent or empty value records
            the comment anonymously. An over-long value yields a 400.
        comment:
          type: string
          description: >-
            Comment body. An empty or whitespace-only value yields a 400.
    CreateCommentResponse:
      type: object
      required:
        - comment_id
      properties:
        comment_id:
          type: string
          description: Identifier generated for the recorded comment.
...